validate graphql schemas
