+++ internal/request/graphql/schema/collection.go
@@ -12,6 +12,7 @@ package schema
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"sort"
 	"strings"
@@ -58,16 +59,23 @@ func FromString(ctx context.Context, schemaString string) (
 	source := source.NewSource(&source.Source{
 		Body: []byte(schemaString),
 	})
-
-	doc, err := gqlp.Parse(
-		gqlp.ParseParams{
-			Source: source,
-		},
-	)
+	doc, err := gqlp.Parse(gqlp.ParseParams{
+		Source: source,
+	})
 	if err != nil {
 		return nil, err
 	}
-
+	schema, err := defaultSchema()
+	if err != nil {
+		return nil, err
+	}
+	validation := gql.ValidateDocument(&schema, doc, gql.SpecifiedRules)
+	if !validation.IsValid {
+		for _, e := range validation.Errors {
+			err = errors.Join(err, e)
+		}
+		return nil, err
+	}
 	return fromAst(doc)
 }
